Further splitted configuration and application state Reorganized configuration modules
